свел все в одну модель, отредактировал заглушку лоадинга
