+++ src/main.js
@@ -2,8 +2,6 @@ import FilterPagePresenter from './presenter/filter-presenter';
 import MainPagePresenter from './presenter/main-presenter';
 import TripNewView from './view/trip-new-view';
 import PointsModel from './model/points-model';
-import OffersModel from './model/offers-model';
-import DestinationsModel from './model/destinations-model';
 import FilterModel from './model/filter-model';
 import PointsApiService from './points-api-service';
 
@@ -15,19 +13,11 @@ const pageMainTripEventElement = document.querySelector('.trip-events');
 const pointsModel = new PointsModel({
   pointsApiService: new PointsApiService(END_POINT, AUTHORIZATION)
 });
-const offersModel = new OffersModel({
-  pointsApiService: new PointsApiService(END_POINT, AUTHORIZATION)
-});
-const destinationsModel = new DestinationsModel({
-  pointsApiService: new PointsApiService(END_POINT, AUTHORIZATION)
-});
 const filterModel = new FilterModel();
 
 const mainPagePresenter = new MainPagePresenter({
   pointsContainer: pageMainTripEventElement,
   pointsModel,
-  offersModel,
-  destinationsModel,
   filterModel,
   onNewPointDestroy: handleNewPointFormClose,
 });
@@ -54,8 +44,3 @@ function handleNewPointButtonClick() {
 filterPagePresenter.init();
 mainPagePresenter.init();
 pointsModel.init();
-// .finally(() => {
-//   newPointButtonComponent.element.disabled = true;
-// });
-offersModel.init();
-destinationsModel.init();
