--- src/model/offers-model.js
@@ -1,28 +1,0 @@
-import Observable from '../framework/observable';
-
-export default class OffersModel extends Observable {
-  #offers = [];
-  #pointsApiService = null;
-
-  constructor({pointsApiService}) {
-    super();
-    this.#pointsApiService = pointsApiService;
-  }
-
-  async init() {
-    try {
-      const offers = await this.#pointsApiService.offers;
-      this.#offers = offers;
-    } catch(err) {
-      this.#offers = [];
-    }
-  }
-
-  get offers() {
-    return this.#offers;
-  }
-
-  set offers (offers) {
-    this.#offers = offers;
-  }
-}
